+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/AnalyzeActionsResult.cs
@@ -17,7 +17,7 @@ namespace Azure.AI.TextAnalytics
             IReadOnlyCollection<RecognizeLinkedEntitiesActionResult> recognizeLinkedEntitiesActionResults,
             IReadOnlyCollection<AnalyzeSentimentActionResult> analyzeSentimentActionResults,
             IReadOnlyCollection<ExtractSummaryActionResult> extractSummaryActionResults,
-            IReadOnlyCollection<ClassifyCustomCategoriesActionResult> classifyCustomCategoriesActionResults
+            IReadOnlyCollection<MultiCategoryClassifyActionResult> multiCategoryClassifyActionResults
             )
         {
             ExtractKeyPhrasesResults = extractKeyPhrasesActionResults;
@@ -26,7 +26,7 @@ namespace Azure.AI.TextAnalytics
             RecognizeLinkedEntitiesResults = recognizeLinkedEntitiesActionResults;
             AnalyzeSentimentResults = analyzeSentimentActionResults;
             ExtractSummaryResults = extractSummaryActionResults;
-            ClassifyCustomCategoriesResults = classifyCustomCategoriesActionResults;
+            MultiCategoryClassifyResults = multiCategoryClassifyActionResults;
         }
 
         /// <summary>
@@ -62,6 +62,6 @@ namespace Azure.AI.TextAnalytics
         /// <summary>
         /// Determines the collection of <see cref="ExtractSummaryActionResult"/>.
         /// </summary>
-        public IReadOnlyCollection<ClassifyCustomCategoriesActionResult> ClassifyCustomCategoriesResults { get; }
+        public IReadOnlyCollection<MultiCategoryClassifyActionResult> MultiCategoryClassifyResults { get; }
     }
 }