+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/TextAnalyticsClient.cs
@@ -2474,9 +2474,9 @@ namespace Azure.AI.TextAnalytics
             {
                 tasks.ExtractiveSummarizationTasks = Transforms.ConvertFromExtractSummaryActionsToTasks(actions.ExtractSummaryActions);
             }
-            if (actions.ClassifyCustomCategoriesActions != null)
+            if (actions.MultiCategoryClassifyActions != null)
             {
-                tasks.CustomMultiClassificationTasks = Transforms.ConvertFromClassifyCustomCategoriesActionsToTasks(actions.ClassifyCustomCategoriesActions);
+                tasks.CustomMultiClassificationTasks = Transforms.ConvertFromMultiCategoryClassifyActionsToTasks(actions.MultiCategoryClassifyActions);
             }
             return tasks;
         }
@@ -2488,7 +2488,7 @@ namespace Azure.AI.TextAnalytics
                 actions.RecognizeLinkedEntitiesActions?.Count > 1 ||
                 actions.ExtractKeyPhrasesActions?.Count > 1 ||
                 actions.AnalyzeSentimentActions?.Count > 1 ||
-                actions.ClassifyCustomCategoriesActions?.Count > 1 ||
+                actions.MultiCategoryClassifyActions?.Count > 1 ||
                 actions.ExtractSummaryActions?.Count > 1)
             {
                 throw new ArgumentException("Multiple of the same action is not currently supported.");